fix: Imports for the ConfluenceLoader:process_page

+++ libs/langchain/langchain/document_loaders/confluence.py
@@ -431,21 +431,30 @@ class ConfluenceLoader(BaseLoader):
         ocr_languages: Optional[str] = None,
         keep_markdown_format: Optional[bool] = False,
     ) -> Document:
-        if keep_markdown_format:
+        if include_comments and keep_markdown_format:
             try:
+                from bs4 import BeautifulSoup  # type: ignore
                 from markdownify import markdownify
             except ImportError:
                 raise ImportError(
-                    "`markdownify` package not found, please run "
-                    "`pip install markdownify`"
+                    "`beautifulsoup4` and  `markdownify` packages not found, please run "
+                    "`pip install beautifulsoup4 markdownify`"
                 )
-        else:
+        elif include_comments or not keep_markdown_format:
             try:
                 from bs4 import BeautifulSoup  # type: ignore
             except ImportError:
                 raise ImportError(
                     "`beautifulsoup4` package not found, please run "
                     "`pip install beautifulsoup4`"
+                )
+        else:
+            try:
+                from markdownify import markdownify
+            except ImportError:
+                raise ImportError(
+                    "`markdownify` package not found, please run "
+                    "`pip install markdownify`"
                 )
 
         if include_attachments:
